youtube-music-bot/requirements.txt
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
Pillow>=10.0.0
pydub>=0.25.0
requests>=2.31.0
python-dotenv>=1.0.0
pytrends>=4.9.2
anthropic>=0.40.0
pytz>=2024.1
scipy>=1.11.0
